Make magic mechanism extensible
Before, it relied on a fixed list of types, which was just not
extensible at all, but it saved some typing when checking magic, since
the compiler could figure out the type.

Now magic numbers are calculated by hashing the name of the explicitly
given type. It's more typing, but you don't need to adjust a big table
of magic number values anymore whenever a new type is added. And of
course you can use this mechanism even with types defined outside of
rouse itself, which is much, much better.

This also removes several of the magic macros to simplify things and
adds magic to some types that were somehow missing it.