Throw out GLEW, load GL stuff manually with SDL
GLEW really doesn't work properly, it reports extensions that aren't
actually supported on Linux and is a nightmare to get working on
Windows, where it doesn't seem to interact with ANGLE properly.

Now we just load all the GL functions manually with
`SDL_GL_GetProcAddress` and check for extensions with
`SDL_GL_ExtensionSupported`.