Add tracking of another sprite's world matrix
A sprite can now force its own world matrix to another sprite's world
matrix instead of using its own transform. This simplifies thwarting the
hierarchy, making a sprite appear with a transform as if it were the
child of a different sprite.

This feature isn't terribly solid at this point, you can very easily
construct constellations where it just doesn't work, but for my use case
it's an ideal fit.