Add clipboard functions and unsafe lrouse build
Makes all those dummy registry index pointers extern so that they're
accessible from outside. Adds a post init callback to fiddle with these
functions and disable them.

With all that, it turns off the clipboard functions in the regular
lrouse build, but leaves them enabled in a new
lrouse-with-questionable-features build, which also allows stuff like
writing to files.