Rework ImGui integration, remove ImNodes
Adds the ImGui bindings generator and the scripting to fetch, build and
generate. Updates ImGui to the latest version. The bindings now look
different, e.g. functions use the same capitalization as in ImGui
instead of turning everything into underscores, it was just too
confusing having to keep translating that from the documentation.

ImNodes is gone for now, don't have time to update those bindings too
and it hasn't been useful.